Asset
The models for Qbert and the enemy have both been given an idle jump animation, which has been tested and can be imported into Unreal. The enemy model is quite well made, and upon import, it requires the scale values for x, y, and z to be increased to 100, and the animation uses the 'Add Animation Asset' to achieve the jump effect. For some reason, Qbert becomes slightly deformed after being imported into Unreal, but this can be corrected with scaling adjustments. In the animation sequence, there is an option that can realize an idle jump, and some preceding unknow animation sequences are ineffective.